+++ Quiz-Website/src/main/java/com/freeuni/proj_100/quizwebsite/service/QuizService.java
@@ -25,7 +25,7 @@ public class QuizService {
     /**
      * Gets a quiz by its ID.
      */
-    public Optional<Quiz> getQuizById(int id) {
+    public Optional<Quiz> getQuizById(Long id) {
         return quizRepository.findById(id);
     }
 
@@ -39,7 +39,7 @@ public class QuizService {
     /**
      * Gets all quizzes created by a specific user.
      */
-    public List<Quiz> getQuizzesByCreator(int creatorId) {
+    public List<Quiz> getQuizzesByCreator(Long creatorId) {
         return quizRepository.findByCreatorId(creatorId);
     }
 }
